requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   starlette
    #   watchfiles
certifi==2026.1.4
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via uvicorn
fastapi==0.129.0
    # via -r requirements.in
geoalchemy2==0.14.4
    # via -r requirements.in
greenlet==3.3.1
    # via sqlalchemy
h11==0.16.0
    # via uvicorn
httptools==0.7.1
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   requests
jinja2==3.1.6
    # via -r requirements.in
markupsafe==3.0.3
    # via jinja2
packaging==26.0
    # via geoalchemy2
psycopg2-binary==2.9.10
    # via -r requirements.in
pydantic==2.12.5
    # via fastapi
pydantic-core==2.41.5
    # via pydantic
python-dotenv==1.0.0
    # via
    #   -r requirements.in
    #   uvicorn
python-multipart==0.0.22
    # via -r requirements.in
pyyaml==6.0.3
    # via uvicorn
requests==2.32.4
    # via -r requirements.in
sqlalchemy==2.0.24
    # via
    #   -r requirements.in
    #   geoalchemy2
starlette==0.52.1
    # via
    #   -r requirements.in
    #   fastapi
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
urllib3==2.6.3
    # via requests
uvicorn[standard]==0.40.0
    # via -r requirements.in
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
